completed till payment

+++ controllers/admin.js
@@ -1,30 +1,28 @@
-const Product = require('../models/product');
+const mongoose = require('mongoose');
+
+const fileHelper = require('../util/file');
+
 const { validationResult } = require('express-validator/check');
 
-exports.getAddProduct = (req, res, next) => {
-  if (!req.session.isLoggedIn) {
-    return res.redirect('/login');
-  }
+const Product = require('../models/product');
 
+exports.getAddProduct = (req, res, next) => {
   res.render('admin/edit-product', {
     pageTitle: 'Add Product',
     path: '/admin/add-product',
     editing: false,
-    isAuthenticated: req.session.isLoggedIn,
     hasError: false,
     errorMessage: null,
-    validationError: []
+    validationErrors: []
   });
 };
 
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+  const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
-  const errors = validationResult(req);
-  console.log(errors.array());
-  if (!errors.isEmpty()) {
+  if (!image) {
     return res.status(422).render('admin/edit-product', {
       pageTitle: 'Add Product',
       path: '/admin/add-product',
@@ -32,17 +30,36 @@ exports.postAddProduct = (req, res, next) => {
       hasError: true,
       product: {
         title: title,
-        imageUrl: imageUrl,
         price: price,
         description: description
       },
-      isAuthenticated: req.session.isLoggedIn,
+      errorMessage: 'Attached file is not an image.',
+      validationErrors: []
+    });
+  }
+  const errors = validationResult(req);
+
+  if (!errors.isEmpty()) {
+    console.log(errors.array());
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/add-product',
+      editing: false,
+      hasError: true,
+      product: {
+        title: title,
+        price: price,
+        description: description
+      },
       errorMessage: errors.array()[0].msg,
-      validationError: errors.array()
+      validationErrors: errors.array()
     });
   }
 
+  const imageUrl = image.path;
+
   const product = new Product({
+    // _id: new mongoose.Types.ObjectId('5badf72403fd8b5be0366e81'),
     title: title,
     price: price,
     description: description,
@@ -57,7 +74,6 @@ exports.postAddProduct = (req, res, next) => {
       res.redirect('/admin/products');
     })
     .catch(err => {
-      // console.log(err);
       // return res.status(500).render('admin/edit-product', {
       //   pageTitle: 'Add Product',
       //   path: '/admin/add-product',
@@ -69,9 +85,8 @@ exports.postAddProduct = (req, res, next) => {
       //     price: price,
       //     description: description
       //   },
-      //   isAuthenticated: req.session.isLoggedIn,
-      //   errorMessage: 'Database Operation failed!',
-      //   validationError: []
+      //   errorMessage: 'Database operation failed, please try again.',
+      //   validationErrors: []
       // });
       // res.redirect('/500');
       const error = new Error(err);
@@ -91,24 +106,20 @@ exports.getEditProduct = (req, res, next) => {
       if (!product) {
         return res.redirect('/');
       }
-      if (product.userId.toString() !== req.user._id.toString()) {
-        return res.redirect('/');
-      }
       res.render('admin/edit-product', {
         pageTitle: 'Edit Product',
         path: '/admin/edit-product',
         editing: editMode,
         product: product,
-        isAuthenticated: req.session.isLoggedIn,
         hasError: false,
-        errorMessage: false,
-        validationError: []
+        errorMessage: null,
+        validationErrors: []
       });
     })
     .catch(err => {
       const error = new Error(err);
       error.httpStatusCode = 500;
-      return next(500);
+      return next(error);
     });
 };
 
@@ -116,9 +127,11 @@ exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId;
   const updatedTitle = req.body.title;
   const updatedPrice = req.body.price;
-  const updatedImageUrl = req.body.imageUrl;
+  const image = req.file;
   const updatedDesc = req.body.description;
+
   const errors = validationResult(req);
+
   if (!errors.isEmpty()) {
     return res.status(422).render('admin/edit-product', {
       pageTitle: 'Edit Product',
@@ -127,14 +140,12 @@ exports.postEditProduct = (req, res, next) => {
       hasError: true,
       product: {
         title: updatedTitle,
-        imageUrl: updatedImageUrl,
         price: updatedPrice,
         description: updatedDesc,
         _id: prodId
       },
-      isAuthenticated: req.session.isLoggedIn,
       errorMessage: errors.array()[0].msg,
-      validationError: errors.array()
+      validationErrors: errors.array()
     });
   }
 
@@ -146,57 +157,56 @@ exports.postEditProduct = (req, res, next) => {
       product.title = updatedTitle;
       product.price = updatedPrice;
       product.description = updatedDesc;
-      product.imageUrl = updatedImageUrl;
+      if (image) {
+        fileHelper.deleteFile(product.imageUrl);
+        product.imageUrl = image.path;
+      }
       return product.save().then(result => {
         console.log('UPDATED PRODUCT!');
         res.redirect('/admin/products');
-      })
-      .catch(err => {
-        const error = new Error(err);
-        error.httpStatusCode = 500;
-        return next(500);
       });
     })
     .catch(err => {
       const error = new Error(err);
       error.httpStatusCode = 500;
-      return next(500);
+      return next(error);
     });
 };
 
 exports.getProducts = (req, res, next) => {
-  Product.find({userId: req.user._id})
+  Product.find({ userId: req.user._id })
     // .select('title price -_id')
     // .populate('userId', 'name')
     .then(products => {
-      // console.log(products);
+      console.log(products);
       res.render('admin/products', {
         prods: products,
         pageTitle: 'Admin Products',
-        path: '/admin/products',
-        isAuthenticated: req.session.isLoggedIn
+        path: '/admin/products'
       });
     })
     .catch(err => {
       const error = new Error(err);
       error.httpStatusCode = 500;
-      return next(500);
+      return next(error);
     });
 };
 
-exports.postDeleteProduct = (req, res, next) => {
-  const prodId = req.body.productId;
-  
-  console.log(prodId);
-  console.log(req.user._id)
-  return Product.deleteOne({_id: prodId, userId: req.user._id})
+exports.deleteProduct = (req, res, next) => {
+  const prodId = req.params.productId;
+  Product.findById(prodId)
+    .then(product => {
+      if (!product) {
+        return next(new Error('Product not found.'));
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
     .then(() => {
       console.log('DESTROYED PRODUCT');
-      res.redirect('/admin/products');
+      res.status(200).json({ message: 'Success!' });
     })
     .catch(err => {
-      const error = new Error(err);
-      error.httpStatusCode = 500;
-      return next(500);
+      res.status(500).json({ message: 'Deleting product failed.' });
     });
 };
